Modify getPixel
We pass in the Surface so that we can calculate the byte offset to the
pixel data based on the surface pixel format.

We have to calculate pitch based on the image width and pixel format
because the pitch field from the SDL_Surface struct is not exposed in
the wrapper.

See: https://github.com/haskell-game/sdl2/issues/175